# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Seedling API",
    "version": "0.1.0",
    "description": "Seedling-HQ MVP API"
  },
...
  "paths": {
...
    "/v1/quotes": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "tenantId": {
                            "type": "string"
                          },
                          "requestId": {
                            "type": "string",
                            "nullable": true
                          },
                          "clientId": {
                            "type": "string"
                          },
                          "propertyId": {
                            "type": "string",
                            "nullable": true
                          },
                          "title": {
                            "type": "string"
                          },
                          "lineItems": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "serviceItemId": {
                                  "type": "string",
                                  "format": "uuid",
                                  "nullable": true
                                },
                                "description": {
                                  "type": "string"
                                },
                                "quantity": {
                                  "type": "number"
                                },
                                "unitPrice": {
                                  "type": "integer"
                                },
                                "total": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "description",
                                "quantity",
                                "unitPrice",
                                "total"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "subtotal": {
                            "type": "number"
                          },
                          "tax": {
                            "type": "number"
                          },
                          "total": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          },
                          "sentAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "approvedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "declinedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "scheduledAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "tenantId",
                          "requestId",
                          "clientId",
                          "propertyId",
                          "title",
                          "lineItems",
                          "subtotal",
                          "tax",
                          "total",
                          "status",
                          "sentAt",
                          "approvedAt",
                          "declinedAt",
                          "scheduledAt",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "cursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "cursor",
                    "hasMore"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "propertyId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "clientId",
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string",
                      "nullable": true
                    },
                    "clientId": {
                      "type": "string"
                    },
                    "propertyId": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "lineItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "serviceItemId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "description": {
                            "type": "string"
                          },
                          "quantity": {
                            "type": "number"
                          },
                          "unitPrice": {
                            "type": "integer"
                          },
                          "total": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "description",
                          "quantity",
                          "unitPrice",
                          "total"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "subtotal": {
                      "type": "number"
                    },
                    "tax": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "sentAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "approvedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "declinedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "scheduledAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "tenantId",
                    "requestId",
                    "clientId",
                    "propertyId",
                    "title",
                    "lineItems",
                    "subtotal",
                    "tax",
                    "total",
                    "status",
                    "sentAt",
                    "approvedAt",
                    "declinedAt",
                    "scheduledAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/quotes/count": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "count"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/quotes/{id}/send": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "not": {}
                  },
                  {
                    "type": "object",
                    "properties": {
                      "expiresInDays": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 90
                      }
                    },
                    "additionalProperties": false,
                    "nullable": true
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quote": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string",
                          "nullable": true
                        },
                        "clientId": {
                          "type": "string"
                        },
                        "propertyId": {
                          "type": "string",
                          "nullable": true
                        },
                        "title": {
                          "type": "string"
                        },
                        "lineItems": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "serviceItemId": {
                                "type": "string",
                                "format": "uuid",
                                "nullable": true
                              },
                              "description": {
                                "type": "string"
                              },
                              "quantity": {
                                "type": "number"
                              },
                              "unitPrice": {
                                "type": "integer"
                              },
                              "total": {
                                "type": "integer"
                              }
                            },
                            "required": [
                              "description",
                              "quantity",
                              "unitPrice",
                              "total"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "subtotal": {
                          "type": "number"
                        },
                        "tax": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string"
                        },
                        "sentAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "approvedAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "declinedAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "scheduledAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "tenantId",
                        "requestId",
                        "clientId",
                        "propertyId",
                        "title",
                        "lineItems",
                        "subtotal",
                        "tax",
                        "total",
                        "status",
                        "sentAt",
                        "approvedAt",
                        "declinedAt",
                        "scheduledAt",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "token": {
                      "type": "string"
                    },
                    "link": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "quote",
                    "token",
                    "link"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}